% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bajranger-package.R
\docType{package}
\name{bajranger-package}
\alias{bajranger}
\alias{bajranger-package}
\title{bajranger: Demultiplexing, Fixing, & Otherwise Futzing With Gnarly Sequencing Data}
\description{
I built a bunch of tools to basically make demultiplexing your data and tried to put it all in one place. Think of this package as the diligent monkey's Shakespeare.
}
\keyword{internal}
