% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elogit.R
\name{elogit}
\alias{elogit}
\title{Empirical logits}
\usage{
elogit(p, N)
}
\arguments{
\item{p}{Proportion of outcome.}

\item{N}{Total number of observations}
}
\description{
Converts proportion into empirical logits (elogit) and corresponding weights (elogit_weights) for regression analysis.
}
\examples{
elogit(.5, 12)
}
\keyword{empirical}
\keyword{logit}
\keyword{proportion,}
