\name{jaeger-package}
\alias{jaeger-package}
\alias{jaeger}
\docType{package}
\title{\packageTitle{jaeger}}
\description{\packageDescription{jaeger}}
\section{Package Content}{\packageIndices{jaeger}}
\author{\packageAuthor{jaeger}}
\section{Maintainer}{\packageMaintainer{jaeger}}
\keyword{package}
