% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rau.R
\name{rau}
\alias{rau}
\title{Rational arcsine transformed units}
\usage{
rau(p, N)
}
\arguments{
\item{p}{Proportion of outcome.}

\item{N}{Total number of observations}
}
\description{
Converts proportion into arcsine transformed units.
}
\examples{
rau(.5, 12)
}
\keyword{arcsine,}
\keyword{proportion,}
\keyword{rational}
